requirements.txt
python-telegram-bot>=13.0
openai>=1.0.0
google-auth>=2.0.0
google-api-python-client>=2.0.0
python-dotenv>=0.19.0
requests>=2.25.0
Pillow>=9.0.0
pillow-heif>=0.10.0
pdf2image>=1.16.0
